+++ libraries/resource_plex_home_theater_service.rb
@@ -1,7 +1,7 @@
 # Encoding: UTF-8
 #
 # Cookbook Name:: plex-home-theater
-# Resource:: plex_home_theater_app
+# Library:: plex_home_theater_service
 #
 # Copyright 2015 Jonathan Hartman
 #
@@ -22,15 +22,13 @@ require 'chef/resource'
 
 class Chef
   class Resource
-    # A parent custom resource for Plex Home Theater App.
+    # A parent custom resource for Plex Home Theater service actions.
     #
     # @author Jonathan Hartman <j@p4nt5.com>
-    class PlexHomeTheaterApp < Resource
-      property :source, [String, nil], default: nil
+    class PlexHomeTheaterService < Resource
+      default_action :nothing
 
-      default_action :install
-
-      %i(install remove).each do |a|
+      %i(enable disable start stop).each do |a|
         action a do
           fail(NotImplementedError,
                "Action '#{a}' must be implemented for '#{self.class}' resource")