+++ libraries/resource_plex_home_theater_service_mac_os_x.rb
@@ -1,7 +1,7 @@
 # Encoding: UTF-8
 #
 # Cookbook Name:: plex-home-theater
-# Resource:: plex_home_theater_service_mac_os_x
+# Library:: plex_home_theater_service_mac_os_x
 #
 # Copyright 2015 Jonathan Hartman
 #
@@ -18,8 +18,9 @@
 # limitations under the License.
 #
 
-require 'mixlib/shell_out'
-require_relative 'plex_home_theater_service'
+require 'mixlib/shellout'
+require_relative 'resource_plex_home_theater_app_mac_os_x'
+require_relative 'resource_plex_home_theater_service'
 
 class Chef
   class Resource
@@ -27,6 +28,8 @@ class Chef
     #
     # @author Jonathan Hartman <j@p4nt5.com>
     class PlexHomeTheaterServiceMacOsX < PlexHomeTheaterService
+      PATH ||= PlexHomeTheaterAppMacOsX::PATH
+
       provides :plex_home_theater_service, platform_family: 'mac_os_x'
 
       #
@@ -41,7 +44,7 @@ class Chef
               'new login item at end with properties ' \
               "{name: \"Plex Home Theater\", path: \"#{PATH}\", " \
               "hidden: false}'"
-        execute 'enable Plex Home Theater' do
+        execute 'Enable Plex Home Theater' do
           command cmd
           action :run
           only_if { !enabled? }
@@ -54,7 +57,7 @@ class Chef
       action :disable do
         cmd = 'osascript -e \'tell application "System Events" to delete ' \
               'login item "Plex Home Theater"\''
-        execute 'disable Plex Home Theater' do
+        execute 'Disable Plex Home Theater' do
           command cmd
           action :run
           only_if { enabled? }
@@ -65,7 +68,7 @@ class Chef
       # Use the `open` command to start up Plex Home Theater.
       #
       action :start do
-        execute 'start Plex Home Theater' do
+        execute 'Start Plex Home Theater' do
           command "open '#{PATH}'"
           user Etc.getlogin
           action :run
@@ -76,7 +79,17 @@ class Chef
         end
       end
 
-      private
+      #
+      # Use `killall` to stop Plex Home Theater.
+      #
+      action :stop do
+        execute 'Stop Plex Home Theater' do
+          command 'killall Plex\\ Home\\ Theater'
+          user Etc.getlogin
+          ignore_failure true
+          action :run
+        end
+      end
 
       #
       # Shell out and use AppleScript to check whether the login item