+++ spec/support/cookbooks/plex_home_theater_app_test/recipes/remove.rb
@@ -1,0 +1,5 @@
+# Encoding: UTF-8
+
+plex_home_theater_app 'default' do
+  action :remove
+end