+++ spec/support/cookbooks/plex_home_theater_service_test/recipes/start.rb
@@ -1,0 +1,5 @@
+# Encoding: UTF-8
+
+plex_home_theater_service 'default' do
+  action :start
+end